refactor(jazz-tools): Phase 1 — relocate junk-drawer modules
Pure mechanical moves (no functional change). Each top-level file moves
into the subsystem that owns it; lib.rs gains re-exports so the existing
crate-public import paths (jazz_tools::sync_tracer, ::binding_support,
::routes) continue to resolve.

- monotonic_clock.rs       -> sync_manager/clock.rs
- sync_tracer.rs           -> sync_manager/sync_tracer.rs
- routes.rs                -> server/routes.rs
- binding_support.rs       -> query_manager/bindings.rs
- test_row_history.rs      -> test_support.rs (gated cfg(any(test, feature = "test-utils")))

Internal callers using crate::test_row_history::* are updated to
crate::test_support::*; one crate::routes use in server/builder.rs is
updated to super::routes. All other callers continue via lib.rs re-exports.

Spec: specs/todo/c_later/crate_structure_cleanup.md (Phase 1).
