+++ crates/jazz-tools/src/query_manager/rebac_tests.rs
@@ -19,7 +19,7 @@ use crate::sync_manager::{
     ClientId, Destination, InboxEntry, QueryId, RowMetadata, Source, SyncError, SyncManager,
     SyncPayload,
 };
-use crate::test_row_history::{
+use crate::test_support::{
     apply_test_row_batch, create_test_row, load_test_row_metadata, load_test_row_tip_ids,
     seeded_memory_storage,
 };
